refactor: better structure
- split util to dynamicUtil, OpenapiUtil, swaggerUtil
- extract inline function from openapi and swagger
