+++ lib/dynamic.js
@@ -1,8 +1,6 @@
 'use strict'
 
-const { addHook } = require('./util')
-const buildSwagger = require('./swagger')
-const buildOpenapi = require('./openapi')
+const { addHook, resolveSwaggerFunction } = require('./dynamicUtil')
 
 module.exports = function (fastify, opts, done) {
   const { routes, Ref } = addHook(fastify)
@@ -22,15 +20,12 @@ module.exports = function (fastify, opts, done) {
   }
 
   const cache = {
-    swaggerObject: null,
-    swaggerString: null
+    object: null,
+    string: null
   }
 
-  if (Object.keys(opts.openapi).length > 0 && opts.openapi.constructor === Object) {
-    fastify.decorate('swagger', buildOpenapi(opts, routes, Ref, cache, done))
-  } else {
-    fastify.decorate('swagger', buildSwagger(opts, routes, Ref, cache, done))
-  }
+  const swagger = resolveSwaggerFunction(opts, routes, Ref, cache, done)
+  fastify.decorate('swagger', swagger)
 
   done()
 }
