+++ lib/openapi.js
@@ -3,7 +3,8 @@
 const fs = require('fs')
 const path = require('path')
 const yaml = require('js-yaml')
-const { formatParamUrl, plainJsonObjectToSwagger2, swagger2ParametersToOpenapi3, stripBasePathByServers } = require('./util')
+const { formatParamUrl } = require('./dynamicUtil')
+const { getBodyParams, getCommonParams, generateResponse, stripBasePathByServers } = require('./openapiUtil')
 
 module.exports = function (opts, routes, Ref, cache, done) {
   let ref
@@ -140,22 +141,22 @@ module.exports = function (opts, routes, Ref, cache, done) {
         }
 
         if (schema.querystring) {
-          getParams('query', parameters, schema.querystring)
+          getCommonParams('query', parameters, schema.querystring, ref, swaggerObject.components.schemas)
         }
 
         if (schema.body) {
           swaggerMethod.requestBody = {
             content: {}
           }
-          getBodyParams(swaggerMethod.requestBody.content, schema.body, schema.consumes)
+          getBodyParams(swaggerMethod.requestBody.content, schema.body, schema.consumes, ref)
         }
 
         if (schema.params) {
-          getParams('path', parameters, schema.params)
+          getCommonParams('path', parameters, schema.params, ref, swaggerObject.components.schemas)
         }
 
         if (schema.headers) {
-          getParams('header', parameters, schema.headers)
+          getCommonParams('header', parameters, schema.headers, ref, swaggerObject.components.schemas)
         }
 
         if (parameters.length) {
@@ -181,7 +182,7 @@ module.exports = function (opts, routes, Ref, cache, done) {
         }
       }
 
-      swaggerMethod.responses = generateResponse(schema ? schema.response : null)
+      swaggerMethod.responses = generateResponse(schema ? schema.response : null, ref)
 
       swaggerObject.paths[url] = swaggerRoute
     }
@@ -194,55 +195,5 @@ module.exports = function (opts, routes, Ref, cache, done) {
 
     cache.swaggerObject = swaggerObject
     return swaggerObject
-
-    function getBodyParams (parameters, body, consumes) {
-      const bodyResolved = ref.resolve(body)
-
-      if ((Array.isArray(consumes) && consumes.length === 0) || typeof consumes === 'undefined') {
-        consumes = ['application/json']
-      }
-
-      consumes.forEach((consume) => {
-        parameters[consume] = {
-          schema: bodyResolved
-        }
-      })
-    }
-
-    function getParams (container, parameters, query) {
-      const resolved = ref.resolve(query)
-      const add = plainJsonObjectToSwagger2(container, resolved, swaggerObject.components.schemas)
-      add.forEach(_ => parameters.push(swagger2ParametersToOpenapi3(_)))
-    }
-
-    // https://swagger.io/docs/specification/2-0/describing-responses/
-    function generateResponse (fastifyResponseJson) {
-      // if the user does not provided an out schema
-      if (!fastifyResponseJson) {
-        return { 200: { description: 'Default Response' } }
-      }
-
-      const responsesContainer = {}
-
-      Object.keys(fastifyResponseJson).forEach(key => {
-        const rawJsonSchema = fastifyResponseJson[key]
-        const resolved = ref.resolve(rawJsonSchema)
-
-        const content = {
-          'application/json': {}
-        }
-
-        content['application/json'] = {
-          schema: resolved
-        }
-
-        responsesContainer[key] = {
-          content,
-          description: rawJsonSchema.description || 'Default Response'
-        }
-      })
-
-      return responsesContainer
-    }
   }
 }
