+++ lib/swagger.js
@@ -3,7 +3,8 @@
 const fs = require('fs')
 const path = require('path')
 const yaml = require('js-yaml')
-const { formatParamUrl, consumesFormOnly, plainJsonObjectToSwagger2 } = require('./util')
+const { formatParamUrl } = require('./dynamicUtil')
+const { consumesFormOnly, getBodyParams, getCommonParams, generateResponse } = require('./swaggerUtil')
 
 module.exports = function (opts, routes, Ref, cache, done) {
   let ref
@@ -157,23 +158,23 @@ module.exports = function (opts, routes, Ref, cache, done) {
         }
 
         if (schema.querystring) {
-          getParams('query', parameters, schema.querystring)
+          getCommonParams('query', parameters, schema.querystring, ref, swaggerObject.definitions)
         }
 
         if (schema.body) {
           const consumesAllFormOnly =
               consumesFormOnly(schema) || consumesFormOnly(swaggerObject)
           consumesAllFormOnly
-            ? getParams('formData', parameters, schema.body)
-            : getBodyParams(parameters, schema.body)
+            ? getCommonParams('formData', parameters, schema.body, ref, swaggerObject.definitions)
+            : getBodyParams(parameters, schema.body, ref)
         }
 
         if (schema.params) {
-          getParams('path', parameters, schema.params)
+          getCommonParams('path', parameters, schema.params, ref, swaggerObject.definitions)
         }
 
         if (schema.headers) {
-          getParams('header', parameters, schema.headers)
+          getCommonParams('header', parameters, schema.headers, ref, swaggerObject.definitions)
         }
 
         if (parameters.length) {
@@ -195,7 +196,7 @@ module.exports = function (opts, routes, Ref, cache, done) {
         }
       }
 
-      swaggerMethod.responses = generateResponse(schema ? schema.response : null)
+      swaggerMethod.responses = generateResponse(schema ? schema.response : null, ref)
 
       swaggerObject.paths[url] = swaggerRoute
     }
@@ -208,45 +209,5 @@ module.exports = function (opts, routes, Ref, cache, done) {
 
     cache.swaggerObject = swaggerObject
     return swaggerObject
-
-    function getBodyParams (parameters, body) {
-      const bodyResolved = ref.resolve(body)
-
-      const param = {}
-      param.name = 'body'
-      param.in = 'body'
-      param.schema = bodyResolved
-      parameters.push(param)
-    }
-
-    function getParams (container, parameters, form) {
-      const resolved = ref.resolve(form)
-      const add = plainJsonObjectToSwagger2(container, resolved, swaggerObject.definitions)
-      add.forEach(_ => parameters.push(_))
-    }
-
-    // https://swagger.io/docs/specification/2-0/describing-responses/
-    function generateResponse (fastifyResponseJson) {
-      // if the user does not provided an out schema
-      if (!fastifyResponseJson) {
-        return { 200: { description: 'Default Response' } }
-      }
-
-      const responsesContainer = {}
-
-      Object.keys(fastifyResponseJson).forEach(key => {
-        // 2xx is not supported by swagger
-
-        const rawJsonSchema = fastifyResponseJson[key]
-        const resolved = ref.resolve(rawJsonSchema)
-
-        responsesContainer[key] = {
-          schema: resolved,
-          description: rawJsonSchema.description || 'Default Response'
-        }
-      })
-
-      return responsesContainer
-    }
   }
 }
